+++ figpack/views/__init__.py
@@ -16,5 +16,7 @@ from .Splitter import Splitter
 from .TabLayout import TabLayout
 from .TabLayoutItem import TabLayoutItem
 from .TimeseriesGraph import TimeseriesGraph
+from .VerticalLayout import VerticalLayout
+from .VerticalLayoutItem import VerticalLayoutItem
 
 from .PlotlyExtension import PlotlyFigure